+++ src/core/converter/ImportFileAeroflyCustomMissionsTmcConverter.ts
@@ -1,12 +1,12 @@
-import { AeroflyMainConfigParser } from "./AeroflyMainConfigParser.js";
+import { AeroflyCustomMissionsParser } from "./AeroflyCustomMissionsParser.js";
 import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 
-export class ImportFileAeroflyMcfConverter implements ImportFileConverter {
-  static readonly fileExtension = "mcf";
+export class ImportFileAeroflyCustomMissionsTmcConverter implements ImportFileConverter {
+  static readonly fileExtension = "tmc";
 
   convert(content: string, flightplan: AeroflyFlight): void {
-    const parser = new AeroflyMainConfigParser();
+    const parser = new AeroflyCustomMissionsParser();
     const newFlightplan = parser.parse(content);
     flightplan.aircraft = newFlightplan.aircraft;
     flightplan.clouds = newFlightplan.clouds;