+++ src/core/services/ImportFileFinderService.ts
@@ -5,6 +5,7 @@ import { ImportFileGarminFpl } from "../converter/ImportFileGarminFplConverter.j
 import { ImportFileMsfs } from "../converter/ImportFileMsfsConverter.js";
 import { ImportFileXplaneFms } from "../converter/ImportFileXplaneFmsConverter.js";
 import { ImportFileAeroflyMcfConverter } from "../converter/ImportFileAeroflyMcfConverter.js";
+import { ImportFileAeroflyCustomMissionsTmcConverter } from "../converter/ImportFileAeroflyCustomMissionsTmcConverter.js";
 
 /**
  * Finds local flight plan files
@@ -25,6 +26,7 @@ export class ImportFileFinderService {
     const files = fs.readdirSync(importDirectory);
     const importFiles = files.filter(
       (file: string) =>
+        file.toLowerCase().endsWith(ImportFileAeroflyCustomMissionsTmcConverter.fileExtension) ||
         file.toLowerCase().endsWith(ImportFileAeroflyMcfConverter.fileExtension) ||
         file.toLowerCase().endsWith(ImportFileGarminFpl.fileExtension) ||
         file.toLowerCase().endsWith(ImportFileMsfs.fileExtension) ||