+++ src/core/util/AeroflyFlightHelper.ts
@@ -1,13 +1,17 @@
-import { AeroflyFlight, AeroflyNavRouteOrigin } from "@fboes/aerofly-custom-missions";
-import { Point } from "@fboes/geojson";
+import {
+  AeroflyFlight,
+  AeroflyNavRouteOrigin,
+  AeroflyNavRouteDepartureRunway,
+  AeroflyNavRouteDestinationRunway,
+} from "@fboes/aerofly-custom-missions";
+import { Point, Vector } from "@fboes/geojson";
 
 /**
  * Offer additional properties derived from `AeroflyFlight` classes
  */
 export class AeroflyFlightHelper {
   /**
-   *
-   * @returns in meters
+   * @returns total flight distance in meters
    */
   static getFlightplanDistance(aeroflyFlight: AeroflyFlight): number {
     let currentCoodinates: Point | null = null;
@@ -32,5 +36,22 @@ export class AeroflyFlightHelper {
     return Math.round(
       (aeroflyFlight.navigation.waypoints.find((wp) => wp instanceof AeroflyNavRouteOrigin)?.longitude ?? 0) / 15,
     );
+  }
+
+  /**
+   * @returns the given runway position moved by its length along its direction to the possible runway threshold (instead of its center)
+   */
+  static positionRunwayWaypoint<T extends AeroflyNavRouteDepartureRunway | AeroflyNavRouteDestinationRunway>(
+    waypoint: T,
+  ): T {
+    const direction_degree = waypoint.direction_degree ?? Number(waypoint.identifier.replace(/\D+/g, "")) * 10;
+    const runwayLength = waypoint.runwayLength ?? 1500;
+
+    const coordinates = new Point(waypoint.longitude, waypoint.latitude);
+    const coordinatesNew = coordinates.getPointBy(new Vector(runwayLength / 2, direction_degree + 180));
+
+    waypoint.latitude = coordinatesNew.latitude;
+    waypoint.longitude = coordinatesNew.longitude;
+    return waypoint;
   }
 }