+++ types/core/converter/ImportFileAeroflyMcfConverter.d.ts
@@ -1,6 +1,6 @@
-import { ImportFileXMLConverter } from "./ImportFileConverter.js";
+import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
-export declare class ImportFileAeroflyMcfConverter extends ImportFileXMLConverter {
+export declare class ImportFileAeroflyMcfConverter implements ImportFileConverter {
   static readonly fileExtension = "mcf";
   convert(content: string, flightplan: AeroflyFlight): void;
 }